requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
pydantic==1.8.2
PyJWT==2.3.0
requests==2.26.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
